fix construct manager, view
