+++ Assets/Scripts/Controllers/LoadingManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 2bd0af5f14aa94101a79065805f55a52
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
