+++ Assets/Scripts/Controllers/MenuManager.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 4de1e098f13754048a35d88f072e0485
+guid: bd56e34b70af04526a5ffd759de36531
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
