+++ UserSettings/EditorUserSettings.asset
@@ -15,13 +15,13 @@ EditorUserSettings:
       value: 22424703114646680e0b0227036c6b15050327192f2d283562041c3aecdf073be7e933a7f234362820
       flags: 0
     RecentlyUsedScenePath-3:
-      value: 22424703114646680e0b0227036c6b15050327192f2d283562051232e6e93a3fddd435ece93f7129370bfb25
-      flags: 0
-    RecentlyUsedScenePath-4:
       value: 22424703114646680e0b0227036c6b15050327192f2d28356204183df7df073be7e933a7f234362820
       flags: 0
-    RecentlyUsedScenePath-5:
+    RecentlyUsedScenePath-4:
       value: 22424703114646680e0b0227036c6b15050327192f2d2835621a0d3fe3f33c07d1e433e7e2742a323016f6
+      flags: 0
+    RecentlyUsedScenePath-5:
+      value: 22424703114646680e0b0227036c6b15050327192f2d283562051232e6e93a3fddd435ece93f7129370bfb25
       flags: 0
     vcSharedLogLevel:
       value: 0d5e400f0650
